add test for record method
Here we are breaking the idea of having only one expect statement in
each example. By taking the time to figure out which expectations
can be combined together you save the tear down and set up steps of
multiple specs touching the database. The downside is the when the first
test fails RSpec aborts and doesn't read the second expecation.